Dev (#303)
* Allow miner specified static difficulty via stratum password (d=<diff>) for Ethereum- and Monero-families of pools. Disables vardiff for the miner.
* Solo Payment Scheme - Fixes #112
* Fixed long integer overflow in Ethereum SendTransactionRequest
* WIP
* Wallet lock only if unlocked
* X16r and X16s hash algos
* fixed btcp links (#297)
* Fixed BTCP explorer url/methods and added address link for btcp

+++ src/MiningCore/Blockchain/Ethereum/DaemonRequests/SendTransactionRequest.cs
@@ -18,6 +18,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
+using System.Numerics;
 using MiningCore.Serialization;
 using Newtonsoft.Json;
 
@@ -53,7 +54,7 @@ namespace MiningCore.Blockchain.Ethereum.DaemonRequests
         /// (Optional) Integer of the value send with this transaction
         /// </summary>
         [JsonConverter(typeof(HexToIntegralTypeJsonConverter<ulong>))]
-        public ulong Value { get; set; }
+        public BigInteger Value { get; set; }
 
         /// <summary>
         /// The compiled code of a contract OR the hash of the invoked method signature and encoded parameters.
